FIX:refactored user routes code from normal routes to controller
